move authentication tests to correctly scoped file
AuthenticateAsync_* tests and the CreateAuthenticationService helper
were added to AuthenticationServiceEmailConfirmationTests.cs, but they
test general authentication behavior, not email confirmation. Move them
to a new AuthenticationServiceAuthenticationTests.cs file so each file
has a single, clear responsibility.
